+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/filesystem/FileDBManager.java
@@ -24,28 +24,15 @@ public final class FileDBManager extends FileDB {
      */
     private final FileDBStorage memory;
 
-    /**
-     * IndexedDB is async, we store in memory while we save it. It will delete from memory when it's done.
-     */
-    private FileDB indexedDB;
-
     FileDBManager() {
         TeaApplicationConfiguration config = ((TeaApplication)Gdx.app).getConfig();
         String storagePrefix = config.storagePrefix;
         localStorage = new FileDBStorage(new StoreLocal(storagePrefix));
         memory = new FileDBStorage(new StoreMemory());
-
-        if(config.useIndexedDB || config.useNewExperimentalAssets) {
-            indexedDB = new IndexedDBStorage();
-        }
     }
 
     @Override
     public InputStream read(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.read(file);
-        }
-
         if(memory.exists(file)) {
             return memory.read(file);
         }
@@ -56,19 +43,11 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public byte[] readBytes(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.readBytes(file);
-        }
         return new byte[0];
     }
 
     @Override
     protected void writeInternal(TeaFileHandle file, byte[] data, boolean append, int expectedLength) {
-        if(indexedDB != null) {
-            indexedDB.writeInternal(file, data, append, expectedLength);
-            return;
-        }
-
         // write larger files into memory: up to 16.384kb into local storage (permanent)
         int localStorageMax = 16384;
         if((data.length >= localStorageMax) || (append && (expectedLength >= localStorageMax))) {
@@ -84,10 +63,6 @@ public final class FileDBManager extends FileDB {
 
     @Override
     protected String[] paths(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.paths(file);
-        }
-
         // combine & return the paths of memory & local storage
         String[] pathsMemory = memory.paths(file);
         String[] pathsLocalStorage = localStorage.paths(file);
@@ -99,10 +74,6 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public boolean isDirectory(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.isDirectory(file);
-        }
-
         if(memory.exists(file)) {
             return memory.isDirectory(file);
         }
@@ -113,29 +84,16 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public void mkdirs(TeaFileHandle file) {
-        if(indexedDB != null) {
-            indexedDB.mkdirs(file);
-            return;
-        }
-
         localStorage.mkdirs(file);
     }
 
     @Override
     public boolean exists(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.exists(file);
-        }
-
         return memory.exists(file) || localStorage.exists(file);
     }
 
     @Override
     public boolean delete(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.delete(file);
-        }
-
         if(memory.exists(file)) {
             return memory.delete(file);
         }
@@ -146,18 +104,11 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public boolean deleteDirectory(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.deleteDirectory(file);
-        }
         throw new GdxRuntimeException("Cannot delete directory (missing implementation): " + file);
     }
 
     @Override
     public long length(TeaFileHandle file) {
-        if(indexedDB != null) {
-            return indexedDB.length(file);
-        }
-
         if(memory.exists(file)) {
             return memory.length(file);
         }
@@ -168,11 +119,6 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public void rename(TeaFileHandle source, TeaFileHandle target) {
-        if(indexedDB != null) {
-            indexedDB.rename(source, target);
-            return;
-        }
-
         if(memory.exists(source)) {
             memory.rename(source, target);
         }
@@ -183,9 +129,6 @@ public final class FileDBManager extends FileDB {
 
     @Override
     public String getPath() {
-        if(indexedDB != null) {
-            return indexedDB.getPath();
-        }
         return null;
     }
 }