+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/filesystem/types/InternalDBStorage.java
@@ -1,0 +1,6 @@
+package com.github.xpenatan.gdx.backends.teavm.filesystem.types;
+
+import com.github.xpenatan.gdx.backends.teavm.filesystem.MemoryFileStorage;
+
+public class InternalDBStorage extends MemoryFileStorage {
+}